Python/Ejercicios/Prueba tecnica edunext/02_your_code/requirements.txt
flake8
black
isort
ipython
pytest==6.2.5
pytest-django
pdbpp